requirements.txt
aiosmtpd==1.4.0
astroid==2.5.1
atpublic==2.1.3
attrs==20.3.0
Faker==6.5.0
isort==5.7.0
lazy-object-proxy==1.5.2
ldap3==2.9
mccabe==0.6.1
py4j==0.10.9.2
pyasn1==0.4.8
pylint==2.7.2
python-dateutil==2.8.1
python-ldap-test==0.3.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
paho-mqtt==1.5.1
